Add benchmark generation script for Hugging Face tokenizers
This commit introduces a new script, `generate_benchmarks.py`, which generates reference benchmark corpora using the Hugging Face tokenizers package. The script is designed to mirror .NET benchmark scenarios, allowing for the production of Python parity data on demand. It includes functionality for downloading model assets, building benchmark cases, and generating output in JSON format for various models. The script also supports advanced cases such as pair encoding and truncation scenarios.

+++ src/ErgoX.VecraX.ML.NLP.Tokenizers.HuggingFace/Options/PaddingOptions.cs
@@ -11,8 +11,8 @@ public enum PaddingDirection
 public sealed class PaddingOptions
 {
     public PaddingDirection Direction { get; }
-    public int PadId { get; }
-    public int PadTypeId { get; }
+    public uint PadId { get; }
+    public uint PadTypeId { get; }
     public string PadToken { get; }
     public int? Length { get; }
     public int? PadToMultipleOf { get; }
@@ -46,8 +46,8 @@ public sealed class PaddingOptions
         }
 
         Direction = direction;
-        PadId = padId;
-        PadTypeId = padTypeId;
+    PadId = (uint)padId;
+    PadTypeId = (uint)padTypeId;
         PadToken = padToken;
         Length = length;
         PadToMultipleOf = padToMultipleOf;
